+++ wxdata/source/scwx/util/time.cpp
@@ -10,7 +10,7 @@ namespace scwx
 namespace util
 {
 
-std::chrono::system_clock::time_point TimePoint(uint16_t modifiedJulianDate,
+std::chrono::system_clock::time_point TimePoint(uint32_t modifiedJulianDate,
                                                 uint32_t milliseconds)
 {
    using namespace std::chrono;
